[RIKU-8] refactor: react-router v7 이상부터 권장되는 <Outlet> 사용, TabNaviagtionUI 제거
